requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Django==4.1.2
django-environ==0.9.0
django-paypal==2.0
idna==3.4
lxml==4.9.1
mypy-extensions==0.4.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.6
requests==2.28.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
